Fix final linting warnings
